chore: split the object code into readable chunks
